+++ test/functional/android/android/search_context_test.rb
@@ -51,6 +51,21 @@ class AppiumLibCoreTest
         @driver.find_element :accessibility_id, 'Cloning' # no error
         @driver.back
       end
+
+      def test_viewmatcher
+        skip 'UiAutomator2 does not support viewtag' if @@core.automation_name != :espresso
+        skip_as_appium_version('1.17.0')
+
+        # Uses withText in 'androidx.test.espresso.matcher.ViewMatchers'
+        e = @driver.find_element :view_matcher, {
+          name: 'withText',
+          args: ['Accessibility'],
+          class: 'androidx.test.espresso.matcher.ViewMatchers'
+        }.to_json
+        e.click
+        @driver.find_element :accessibility_id, 'Custom View'
+        @driver.back
+      end
     end
   end
 end